feat: Implement native authentication service with sign-up and password reset flows
- Add request normalizers for sign-up and password reset payloads.
- Create nativeAuthService to handle authentication logic including sign-in, sign-up, and password reset.
- Introduce signUpAttributes service to build sign-up attribute payloads.
- Implement JWT decoding utility for handling authentication tokens.
- Create user utility to build user objects from claims.
- Add validation schemas for sign-in, sign-up, and password reset requests using AJV.
