+++ native-auth-function-app/src/utils/jwt.js
@@ -1,0 +1,12 @@
+export const decodeJwtPayload = (jwt) => {
+    const parts = jwt.split('.');
+    if (parts.length !== 3) {
+        throw new Error('Invalid JWT structure');
+    }
+
+    const base64Url = parts[1];
+    const base64 = base64Url.replace(/-/g, '+').replace(/_/g, '/');
+    const padding = '='.repeat((4 - (base64.length % 4)) % 4);
+    const decoded = Buffer.from(base64 + padding, 'base64').toString();
+    return JSON.parse(decoded);
+};
