+++ native-auth-function-app/src/utils/user.js
@@ -1,0 +1,9 @@
+export const buildUserFromClaims = (claims, fallbackEmail) => ({
+    id: claims?.sub || null,
+    username: claims?.preferred_username || claims?.email || fallbackEmail || null,
+    email: claims?.email || fallbackEmail || claims?.preferred_username || null,
+    firstName: claims?.given_name || '',
+    lastName: claims?.family_name || '',
+    name: claims?.name || claims?.preferred_username || fallbackEmail || '',
+    tenantId: claims?.tid || claims?.tenantId || null
+});
